requirements.txt
DateTime==5.2
prettytable==3.9.0
psycopg2==2.9.7
python-dotenv==1.0.0
